requirements.txt
# FastAPI backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.4.2  # Compatible with Python 3.13
python-multipart==0.0.6

# Machine Learning
tensorflow>=2.17.0
keras>=3.0.0
numpy>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.9.0
